Add rune requirements to spell book teleports
